+++ CHANGELOG.md
@@ -10,6 +10,17 @@
   Therefore, `Trainer.fit` expects an `OperatorDataset` now.
 - A `Criterion` now enables stopping the training loop.
 - The `plotting` module has been removed.
+- Add `timeseries.ipynb` example.
+- Add `Function`, `FunctionSet`, and `FunctionOperatorDataset` classes.
+- Add `function.ipynb` example.
+- Add `Benchmark` base class.
+- Add `SineBenchmark`.
+- Implement `DeepNeuralOperator`.
+- Generalize `NeuralOperator` to take a list of operators.
+- The `data.DatasetShapes` class becomes `operators.OperatorShapes` without `num_observations` attribute.
+- Change `torch` dependency from "==2.1.0" to ">=2.1.0,<3.0.0".
+- Change `dataptation` dependency from "==3.1" to ">=3.1,<4.0".
+- Change `optuna` dependency from "3.5.0" to ">=3.5.0,<4.0.0".
 - Add `FourierLayer` implementation.
 
 ## 0.0.0 (2024-02-22)